+++ lib/_prerequisites-stack.ts
@@ -1,13 +1,9 @@
 import * as cdk from "aws-cdk-lib";
 import { Construct } from "constructs";
 import { Prerequisites } from "./types/config";
-import { aws_s3 as s3 } from "aws-cdk-lib";
-import { aws_kms as kms } from "aws-cdk-lib";
-import { aws_iam as iam } from "aws-cdk-lib";
-import { aws_lambda as lambda } from "aws-cdk-lib";
-import { aws_logs as logs } from "aws-cdk-lib";
-import { aws_fms as fms } from "aws-cdk-lib";
-import { aws_sns as sns } from "aws-cdk-lib";
+import { aws_s3 as s3, aws_kms as kms, aws_iam as iam, aws_lambda as lambda, aws_lambda_nodejs as NodejsFunction, aws_logs as logs  } from "aws-cdk-lib";
+import * as path from "path";
+
 export interface StackProps extends cdk.StackProps {
     readonly prerequisites: Prerequisites;
   }
@@ -31,16 +27,21 @@ export class PrerequisitesStack extends cdk.Stack {
         Messenger="Teams";
         WebhookUrl=props.prerequisites.Information.TeamsWebhook;
       }
-      const ManagedRuleGroupInfo = new lambda.Function(this, "AWS-Firewall-Factory-ManagedRuleGroupInfo", {
-        runtime: lambda.Runtime.PYTHON_3_11,
-        code: lambda.Code.fromAsset("./lib/lambda/ManagedRuleGroupInfo"),
-        handler: "index.lambda_handler",
+      const ManagedRuleGroupInfo = new NodejsFunction.NodejsFunction(this, "AWS-Firewall-Factory-ManagedRuleGroupInfo", {
+        architecture: lambda.Architecture.ARM_64,
+        entry: path.join(__dirname, "../lib/lambda/ManagedRuleGroupInfo/index.ts"),
+        handler: "handler",
         timeout: cdk.Duration.seconds(30),
         environment: {
-          "Messenger": Messenger,
-          "WebhookUrl": WebhookUrl,
+          "MESSENGER": Messenger,
+          "WEBHOOK_URL": WebhookUrl,
         },
         logRetention: logs.RetentionDays.ONE_WEEK,
+        runtime: lambda.Runtime.NODEJS_20_X,
+        memorySize: 128,
+        bundling: {
+          minify: true,
+        },
         description: "Lambda Function to send AWS managed rule group change status notifications (like upcoming new versions and urgent security updates) to messengers (Slack/Teams)",
       });
       ManagedRuleGroupInfo.addToRolePolicy(new iam.PolicyStatement({
@@ -53,47 +54,6 @@ export class PrerequisitesStack extends cdk.Stack {
         sourceArn: "arn:aws:sns:us-east-1:248400274283:aws-managed-waf-rule-notifications",
       });
     }
-    if(props.prerequisites.DdosNotifications) {
-      console.log("📢  Creating Lambda Function that send notifications about potential DDoS activity for protected resources to messengers (Slack/Teams)");
-      let Messenger:string = "";
-      let WebhookUrl:string = "";
-      if(props.prerequisites.DdosNotifications.SlackWebhook) {
-        Messenger="Slack";
-        WebhookUrl=props.prerequisites.DdosNotifications.SlackWebhook;
-      }
-      if(props.prerequisites.DdosNotifications.TeamsWebhook) {
-        Messenger="Teams";
-        WebhookUrl=props.prerequisites.DdosNotifications.TeamsWebhook;
-      }
-      const FmsNotification = new lambda.Function(this, "AWS-Firewall-Factory-FMS-Notifications", {
-        runtime: lambda.Runtime.PYTHON_3_11,
-        code: lambda.Code.fromAsset("./lib/lambda/FmsNotification"),
-        handler: "index.lambda_handler",
-        timeout: cdk.Duration.seconds(30),
-        environment: {
-          "Messenger": Messenger,
-          "WebhookUrl": WebhookUrl,
-        },
-        logRetention: logs.RetentionDays.ONE_WEEK,
-        description: "Lambda Function that send notifications about potential DDoS activity for protected resources to messengers (Slack/Teams)",
-      });
-      const snsRoleName = `arn:aws:iam::${props.env?.account}:role/aws-service-role/fms.amazonaws.com/AWSServiceRoleForFMS`;
-      const FmsTopic = new sns.Topic(this, "FMS-Notifications-Topic");
-      FmsTopic.addToResourcePolicy(new iam.PolicyStatement({
-        actions: ["sns:Publish"],
-        principals: [iam.Role.fromRoleArn(this, "AWSServiceRoleForFMS",snsRoleName)],
-      }));
-      FmsNotification.addPermission("InvokeByFmsSnsTopic", {
-        action: "lambda:InvokeFunction",
-        principal: new iam.ServicePrincipal("sns.amazonaws.com"),
-        sourceArn: FmsTopic.topicArn,
-      });
-      new fms.CfnNotificationChannel(this, "AWS-Firewall-Factory-FMS-NotificationChannel", {
-        snsRoleName,
-        snsTopicArn: FmsTopic.topicArn,
-      });
-    }
-    
 
     if(props.prerequisites.Logging) {
       if(props.prerequisites.Logging.FireHoseKey) {
