+++ lib/lambda/ManagedRuleGroupInfo/messengers/teams/types.ts
@@ -1,0 +1,3 @@
+export interface IncomingWebhookResult {
+    status: number;
+  }
