+++ lib/lambda/SharedComponents/types/index.ts
@@ -1,0 +1,8 @@
+import {ManagedRuleGroupVersion} from "@aws-sdk/client-wafv2";
+
+export interface PaginatedManagedRuleGroupVersions {
+    Versions: ManagedRuleGroupVersion[],
+    CurrentDefaultVersion: string,
+    Error: string,
+}
+
