+++ lib/tools/helpers/pricing/calculator.ts
@@ -7,10 +7,9 @@
 /* eslint-disable @typescript-eslint/no-unsafe-assignment */
 /* eslint-disable @typescript-eslint/no-for-in-array */
 /* eslint-disable @typescript-eslint/no-unsafe-return */
-import { PricingClient, GetProductsCommand, GetProductsCommandInput } from "@aws-sdk/client-pricing";
-import { RuntimeProperties } from "../types/runtimeprops";
-import { Config } from "../types/config";
-import { PriceRegions } from "../types/config";
+import { PricingClient, GetProductsCommand, GetProductsCommandInput, FilterType } from "@aws-sdk/client-pricing";
+import { RuntimeProperties } from "../../../types/runtimeprops";
+import { Config, PriceRegions } from "../../../types/config";
 import {CloudWatchClient, ListDashboardsCommand, ListDashboardsCommandInput } from "@aws-sdk/client-cloudwatch";
 import { ShieldClient, GetSubscriptionStateCommand } from "@aws-sdk/client-shield";
 
@@ -22,7 +21,7 @@ const PRICING_API_ENDPOINT_REGION = "us-east-1";
 
 
 /**
- *
+ * Find Values in Object - Needed for aws pricing api
  * @param obj object where the key is included
  * @param key key which includes the needed value
  * @returns value from key
@@ -31,22 +30,26 @@ function findValues(obj: any, key: string){
   return findValuesHelper(obj, key, []);
 }
 
+/**
+ * Helper function for findValues seachers in a list of Objects for the correct key
+ * @param obj any
+ * @param key string
+ * @param list any
+ * @returns list
+ */
 function findValuesHelper(obj:any, key:string, list: any) {
   if (!obj) return list;
   if (obj instanceof Array) {
-    for (const i in obj) {
+    for (const i of obj) {
       list = list.concat(findValuesHelper(obj[i], key, []));
     }
     return list;
   }
   if (obj[key]) list.push(obj[key]);
 
-  if ((typeof obj === "object") && (obj !== null) ){
-    const children = Object.keys(obj);
-    if (children.length > 0){
-      for (let i = 0; i < children.length; i++ ){
-        list = list.concat(findValuesHelper(obj[children[i]], key, []));
-      }
+  if (typeof obj === "object" && obj !== null) {
+    for (const child of Object.values(obj)) {
+      list = list.concat(findValuesHelper(child, key, []));
     }
   }
   return list;
@@ -59,7 +62,8 @@ function findValuesHelper(obj:any, key:string, list: any) {
  * @param runtimeProps runtime properties object, where to store prices
  * @returns true if prices are update in runtimeprops
  */
-export async function getCurrentPrices(deploymentRegion: PriceRegions, runtimeProps: RuntimeProperties, config: Config, awsregion: string): Promise<boolean> {
+async function getCurrentWafPrices(deploymentRegion: PriceRegions, runtimeProps: RuntimeProperties, config: Config, awsregion: string): Promise<boolean> {
+  console.log("   🔎  Getting current prices for: ", deploymentRegion,"\n");
   try{
     runtimeProps.Pricing.Policy = Number(await getProductPrice(deploymentRegion,"AWSFMS","WAFv2"));
     runtimeProps.Pricing.Rule = Number(await getProductPrice(deploymentRegion,"awswaf",undefined,"Rule"));
@@ -121,28 +125,28 @@ async function getDashboardPrice(deploymentRegion: string, config: Config ): Pro
  */
 async function getProductPrice(deploymentRegion: PriceRegions, servicecode: string, operation?: string,group?: string, groupDescription?: string): Promise<number> {
   const client = new PricingClient({region: PRICING_API_ENDPOINT_REGION});
-  const filters: {Type: string, Field: string, Value: string}[] = [];
+  const filters: {Type: FilterType, Field: string, Value: string}[] = [];
   if(groupDescription){
     filters.push({
-      Type: "TERM_MATCH",
+      Type: FilterType.TERM_MATCH,
       Field: "groupDescription",
       Value: groupDescription});
   }
   if(group){
     filters.push({
-      Type: "TERM_MATCH",
+      Type: FilterType.TERM_MATCH,
       Field: "group",
       Value: group});
   }
   if(operation){
     filters.push({
-      Type: "TERM_MATCH",
+      Type: FilterType.TERM_MATCH,
       Field: "operation",
       Value: operation
     });
   }
   filters.push({
-    Type: "TERM_MATCH",
+    Type: FilterType.TERM_MATCH,
     Field: "location",
     Value: deploymentRegion
   });
@@ -156,7 +160,7 @@ async function getProductPrice(deploymentRegion: PriceRegions, servicecode: stri
   if (!response.PriceList || !response.PriceList[0]) {
     throw new Error("Price list does not exist");
   }
-  const priceList = response.PriceList[0] as any;
+  const priceList = response.PriceList[0];
   const usd = findValues(JSON.parse(priceList.toJSON()),"USD");
   return usd || 0;
 }
@@ -167,32 +171,61 @@ async function getProductPrice(deploymentRegion: PriceRegions, servicecode: stri
  * @param runtimeProps runtime properties object, where to get prices
  * @returns whether price is successfully calculated or not
  */
-export async function isPriceCalculated(runtimeProps: RuntimeProperties): Promise<boolean> {
+export async function isWafPriceCalculated(deploymentRegion: PriceRegions,runtimeProps: RuntimeProperties, config: Config, awsregion: string): Promise<boolean> {
+  const shieldSubscriptionState = await getShieldSubscriptionState();
+  console.log("\n🛡️  Shield Advanced State: " + shieldSubscriptionState);
+
+  console.log("\n💰 Cost: \n");
+
+  await getCurrentWafPrices(deploymentRegion, runtimeProps, config, awsregion);
   const preprocessfixedcost = (runtimeProps.PreProcess.CustomRuleCount * runtimeProps.Pricing.Rule) + runtimeProps.PreProcess.CustomRuleGroupCount + runtimeProps.PreProcess.ManagedRuleGroupCount;
   const postprocessfixedcost = (runtimeProps.PostProcess.CustomRuleCount * runtimeProps.Pricing.Rule) + runtimeProps.PostProcess.CustomRuleGroupCount + runtimeProps.PostProcess.ManagedRuleGroupCount;
   const captchacost = (runtimeProps.PostProcess.CustomCaptchaRuleCount + runtimeProps.PreProcess.CustomCaptchaRuleCount) * runtimeProps.Pricing.Captcha;
   const botcontrolfixedcost = (runtimeProps.PostProcess.ManagedRuleBotControlCount + runtimeProps.PreProcess.ManagedRuleBotControlCount) * runtimeProps.Pricing.BotControl;
   const atpfixedcost =  (runtimeProps.PostProcess.ManagedRuleATPCount + runtimeProps.PreProcess.ManagedRuleATPCount) * runtimeProps.Pricing.AccountTakeoverPrevention;
+
   let fixedcost = runtimeProps.Pricing.Policy + runtimeProps.Pricing.WebACL + postprocessfixedcost + preprocessfixedcost + botcontrolfixedcost  + atpfixedcost + runtimeProps.Pricing.Dashboard;
   const requestscost = runtimeProps.Pricing.Request;
   const totalcost = fixedcost + (requestscost * 5) + (captchacost * 5);
-  const shieldSubscriptionState = await getShieldSubscriptionState();
-  console.log("\n🛡️  Shield Advanced State: " + shieldSubscriptionState);
-  console.log("\n💰 Cost: \n");
-  if(shieldSubscriptionState && shieldSubscriptionState === "ACTIVE"){
+
+  if (shieldSubscriptionState && shieldSubscriptionState === "ACTIVE") {
     fixedcost = botcontrolfixedcost  + atpfixedcost + runtimeProps.Pricing.Dashboard;
   }
+
   console.log("   WAF Rules cost: " + fixedcost + " $ per month");
-  console.log("   WAF Requests: "+ requestscost + " $ pro 1 mio requests");
-  (captchacost > 0) ? console.log("   WAF Analysis fee:\n    Captcha: " +captchacost +"$ per thousand challenge attempts analyzed") : " ";
+  console.log("   WAF Requests: " + requestscost.toFixed(2) + " $ pro 1 mio requests");
+
+  if (captchacost > 0) {
+    console.log("   WAF Analysis fee:\n    Captcha: " + captchacost + "$ per thousand challenge attempts analyzed");
+  } else {
+    console.log(" ");
+  }
+
   console.log("\n   Total WAF cost (monthly): "+ totalcost + " $ *");
-  console.log("\n    * This costs are based on expectation that the WAF gets 5 mio requests per month. ");
-  (atpfixedcost !== 0) ? console.log("\n    *This costs are based on expectation that 10.000 login attempts where analyzed. ") : "";
-  console.log("\n   ℹ The costs are calculated based on the provided information at https://aws.amazon.com/waf/pricing/. ");
-  (botcontrolfixedcost !== 0) ? console.log("     The deployed WAF includes BotControl rules this costs an extra fee of "+runtimeProps.Pricing.BotControl +" $ and " +runtimeProps.Pricing.BotControlRequest +"$ pro 1 mio requests (10 mio request Free Tier). \n     These costs are already included in the price calculation.") : "";
-  (atpfixedcost !== 0) ? console.log("     The deployed WAF includes Account Takeover Prevention rules this costs an extra fee of "+runtimeProps.Pricing.AccountTakeoverPrevention+" $ and " + runtimeProps.Pricing.AccountTakeoverPreventionRequest +" $ per thousand login attempts analyzed (10,000 attempts analyzed Free Tier). \n     These costs are already included in the price calculation.") : "";
-  (runtimeProps.Pricing.Dashboard !== 0) ? console.log("     The deployed WAF includes CloudWatch Dashboard and you have more than 3 Dashboards (Free tier), so you will need to pay " + runtimeProps.Pricing.Dashboard+ "$ for this CloudWatch Dashboard. \n     These costs are already included in the price calculation.") : "";
-  (shieldSubscriptionState === "ACTIVE") ? console.log("     AWS WAF WebACLs or Rules created by Firewall Manager - are Included in AWS Shield Advanced. More information at https://aws.amazon.com/firewall-manager/pricing/.") : "";
+  console.log("\n    * This costs are based on the expectation that the WAF gets 5 mio requests per month. ");
+
+  if (atpfixedcost !== 0) {
+    console.log("\n    * This costs are based on the expectation that 10.000 login attempts were analyzed. ");
+  }
+
+  console.log("\n      ℹ The costs are calculated based on the provided information at https://aws.amazon.com/waf/pricing/. ");
+
+  if (botcontrolfixedcost !== 0) {
+    console.log("     The deployed WAF includes BotControl rules this costs an extra fee of " + runtimeProps.Pricing.BotControl + " $ and " + runtimeProps.Pricing.BotControlRequest + "$ pro 1 mio requests (10 mio request Free Tier). \n     These costs are already included in the price calculation.");
+  }
+
+  if (atpfixedcost !== 0) {
+    console.log("     The deployed WAF includes Account Takeover Prevention rules this costs an extra fee of " + runtimeProps.Pricing.AccountTakeoverPrevention + " $ and " + runtimeProps.Pricing.AccountTakeoverPreventionRequest + " $ per thousand login attempts analyzed (10,000 attempts analyzed Free Tier). \n     These costs are already included in the price calculation.");
+  }
+
+  if (runtimeProps.Pricing.Dashboard !== 0) {
+    console.log("     The deployed WAF includes CloudWatch Dashboard and you have more than 3 Dashboards (Free tier), so you will need to pay " + runtimeProps.Pricing.Dashboard + "$ for this CloudWatch Dashboard. \n     These costs are already included in the price calculation.");
+  }
+
+  if (shieldSubscriptionState === "ACTIVE") {
+    console.log("     AWS WAF WebACLs or Rules created by Firewall Manager - are Included in AWS Shield Advanced. More information at https://aws.amazon.com/firewall-manager/pricing/.");
+  }
+
   console.log("\n\n");
   const pricecalculated = true;
   return pricecalculated;
