+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "aws-firewall-factory",
-  "version": "4.1.6",
+  "version": "4.2.0",
   "bin": {
     "firewallfactory": "bin/aws-firewall-factory.js"
   },
@@ -12,37 +12,41 @@
     "lint": "eslint --ext .ts ."
   },
   "devDependencies": {
-    "@types/node": "20.8.10",
-    "@typescript-eslint/eslint-plugin": "6.13.2",
-    "@typescript-eslint/parser": "6.7.5",
-    "aws-cdk": "2.100.0",
-    "eslint": "8.53.0",
+    "@types/node": "20.11.4",
+    "@typescript-eslint/eslint-plugin": "6.19.0",
+    "@typescript-eslint/parser": "6.19.0",
+    "aws-cdk": "2.121.1",
+    "eslint": "8.56.0",
     "jest": "29.7.0",
     "ts-jest": "29.1.1",
-    "ts-node": "10.9.1",
-    "typescript": "5.2.2"
+    "ts-node": "10.9.2",
+    "typescript": "5.3.3"
   },
   "dependencies": {
-    "@aws-sdk/client-cloudformation": "^3.428.0",
-    "@aws-sdk/client-cloudwatch": "^3.427.0",
-    "@aws-sdk/client-fms": "^3.427.0",
-    "@aws-sdk/client-pricing": "^3.427.0",
-    "@aws-sdk/client-service-quotas": "^3.427.0",
-    "@aws-sdk/client-shield": "^3.428.0",
-    "@aws-sdk/client-wafv2": "^3.427.0",
-    "@babel/traverse": "^7.23.2",
+    "@aws-sdk/client-cloudformation": "^3.490.0",
+    "@aws-sdk/client-cloudwatch": "^3.490.0",
+    "@aws-sdk/client-fms": "^3.490.0",
+    "@aws-sdk/client-pricing": "^3.490.0",
+    "@aws-sdk/client-service-quotas": "^3.490.0",
+    "@aws-sdk/client-shield": "^3.490.0",
+    "@aws-sdk/client-wafv2": "^3.490.0",
+    "@babel/traverse": "^7.23.7",
     "@mhlabs/cfn-diagram": "^1.1.38",
-    "@types/aws-lambda": "^8.10.130",
-    "@types/lodash": "4.14.178",
+    "@slack/types": "^2.11.0",
+    "@slack/webhook": "^7.0.2",
+    "@types/aws-lambda": "^8.10.131",
+    "@types/lodash": "4.14.202",
     "@types/uuid": "^9.0.7",
-    "aws-cdk-lib": "2.100.0",
+    "adaptivecards": "^3.0.2",
+    "aws-cdk-lib": "2.121.1",
     "aws-lambda": "^1.0.7",
+    "axios": "^1.6.5",
     "cfonts": "^3.2.0",
-    "constructs": "10.2.25",
+    "constructs": "10.3.0",
     "lodash": "4.17.21",
     "table": "^6.8.1",
     "typedoc-plugin-extras": "^3.0.0",
-    "typedoc-plugin-keywords": "^1.5.0",
+    "typedoc-plugin-keywords": "^1.6.0",
     "uuid": "^9.0.1"
   }
 }
