+++ tsconfig.json
@@ -3,7 +3,8 @@
     "target": "ES2018",
     "module": "commonjs",
     "lib": [
-      "es2018"
+      "es2018",
+      "dom"
     ],
     "declaration": true,
     "strict": true,
